% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{a_matrix_to_relax_backprop}
\alias{a_matrix_to_relax_backprop}
\title{Back-propagate energy derivative from relaxation rates to a matrix}
\usage{
a_matrix_to_relax_backprop(d_relax_d_a_matrix, d_energy_d_relax)
}
\arguments{
\item{d_relax_d_a_matrix}{matrix (pairs, internal eigenvalues) from \code{gradient} attribute of
\code{a_matrix_to_relax()}}

\item{d_energy_d_relax}{vector (pairs)}
}
\value{
Matrix \verb{(pairs, internal eigenvalues)} with \code{d_energy_d_a_matrix}.
}
\description{
For pair \eqn{p} and internal correlation function component \eqn{j},
the chain rule gives
\deqn{\frac{\partial E}{\partial a_{pj}} =
\frac{\partial E}{\partial R_p}
\frac{\partial R_p}{\partial a_{pj}}.}
}
\seealso{
\code{\link[=a_matrix_to_relax]{a_matrix_to_relax()}}
}
