% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{a_matrix_to_sigma_backprop}
\alias{a_matrix_to_sigma_backprop}
\title{Back-propagate energy derivative from sigma to a matrix}
\usage{
a_matrix_to_sigma_backprop(d_sigma_d_a_matrix, d_energy_d_sigma)
}
\arguments{
\item{d_sigma_d_a_matrix}{matrix (pairs, eigenvalues) fom \code{gradient} attribute of
\code{a_matrix_to_sigma()}}

\item{d_energy_d_sigma}{vector (pairs)}
}
\value{
matrix (pairs, eigenvalues) with d_energy_d_a_matrix
}
\description{
Back-propagate energy derivative from sigma to a matrix
}
