% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{array_shift}
\alias{array_shift}
\title{Shift from one array dimension to another}
\usage{
array_shift(a, n = NULL, dnames = NULL, dfrom = 1, dto = 2)
}
\arguments{
\item{a}{array whose dimensions should be shifted}

\item{n}{integer factor to divide \code{dfrom} dimension and multiply \code{dto} dimension}

\item{dnames}{list with dimension names as alternative to specifying \code{n}}

\item{dfrom}{integer with dimension to move from}

\item{dto}{integer with dimension to move to (should be one greater than \code{dfrom})}
}
\description{
Shift from one array dimension to another
}
