% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{atom_lines_to_coord}
\alias{atom_lines_to_coord}
\title{Convert PDB formatted atom lines into a matrix}
\usage{
atom_lines_to_coord(atom_lines)
}
\arguments{
\item{atom_lines}{character vector of ATOM lines}
}
\value{
3xN matrix with a column for every atom
}
\description{
Generate a matrix containing coordinates in a set of PDB atom lines
}
\details{
The column names in the resulting matrix correspond to columns 13 through 27 of the
ATOM lines. The definitions of the PDB columns and corresponding colnames columns are
as follows:
\tabular{lllll}{
PDB \tab colnames \tab Data Type \tab Field \tab Definition \cr
13-16 \tab 1-4 \tab Atom \tab name \tab Atom name \cr
17 \tab 5 \tab Character \tab altLoc \tab Alternate location indicator \cr
18-20 \tab 6-8 \tab Residue name \tab resName \tab Residue name \cr
22 \tab 10 \tab Character \tab chainID \tab Chain identifier \cr
23-26 \tab 11-4 \tab Integer \tab resSeq \tab Residue sequence number \cr
27 \tab 15 \tab AChar \tab iCode \tab Code for insertion of residues
}

See \url{http://www.wwpdb.org/documentation/file-format-content/format33/sect9.html#ATOM}
for more information.
}
