% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\docType{data}
\name{amber_to_pdb}
\alias{amber_to_pdb}
\alias{pdb_to_amber}
\alias{charmm_to_pdb}
\alias{pdb_to_charmm}
\title{Convert between proton atom/residue name pairs}
\format{
An object of class \code{character} of length 161.

An object of class \code{character} of length 161.

An object of class \code{character} of length 161.

An object of class \code{character} of length 161.
}
\usage{
amber_to_pdb

pdb_to_amber

charmm_to_pdb

pdb_to_charmm
}
\value{
Named character vector mapping atom/residue identifiers between naming
conventions.
}
\description{
Convert between proton atom/residue name pairs
}
\keyword{datasets}
