% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{bond_separation}
\alias{bond_separation}
\title{Calculate bond distances between a set of atoms}
\usage{
bond_separation(coord_mat, depth = 4)
}
\arguments{
\item{depth}{maximum bond distance to calculate}

\item{coord}{3xN matrix with a column for every atom}
}
\value{
NxN matrix with bond distances between atoms (or NA if not calculated)
}
\description{
Calculate bond distances between a set of atoms
}
