% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{calc_subset_mat}
\alias{calc_subset_mat}
\title{Determine subsets of states within transition rate matrices}
\usage{
calc_subset_mat(
  trans_rate_eigen,
  rate_groups = get_rate_groups(-trans_rate_eigen$values[-1])
)
}
\arguments{
\item{trans_rate_eigen}{list with \code{values} and \code{vectors} as returned by \code{eigen()}}

\item{rate_groups}{list with numerically equivalent rates organized into groups

This only works with rate matrices returned by \code{rate_mat_simple()} and
\code{rate_mat_intra_inter}. It does not work with \code{rate_mat_kronecker()}.}
}
\description{
Determine subsets of states within transition rate matrices
}
\examples{
rate_mat <- rate_mat_intra_inter(c(0,0,1,1), 10, 4)
rate_mat <- rate_mat_diag(rate_mat)
calc_subset_mat(eigen(rate_mat))

}
