% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{collapse_array}
\alias{collapse_array}
\title{Collapse first two dimensions of an array}
\usage{
collapse_array(arr, idx_list, func = `+`)
}
\arguments{
\item{arr}{two or three dimensional array with the first two dimensions having the same length}

\item{idx_list}{list of character or integer indices to accumulate using func}

\item{func}{function to apply}
}
\description{
Collapse first two dimensions of an array
}
