% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_array_to_g}
\alias{coord_array_to_g}
\title{Calculate group norm squared values from atomic coordinates}
\usage{
coord_array_to_g(coord_array, atom_pairs, grouping_list)
}
\arguments{
\item{coord_array}{3D array (atoms, xyz, models) with atomic coordinates}

\item{atom_pairs}{matrix with each row having the names or indices of an atom pair (first dimension in \code{coord_array})}

\item{grouping_list}{list of lists of integer vectors giving groupings of models to average interaction tensors}
}
\value{
matrix (pairs, groupings) with group norm squared values
}
\description{
Calculate group norm squared values from atomic coordinates
}
