% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_array_to_g_energy_refactored}
\alias{coord_array_to_g_energy_refactored}
\title{Calculate g value restraint energy from atomic coordinates}
\usage{
coord_array_to_g_energy_refactored(
  coord_array,
  atom_pairs,
  grouping_mat,
  g0,
  loss_func = power_scaled_loss,
  ...,
  gradient = FALSE
)
}
\arguments{
\item{coord_array}{3D array (atoms, xyz, models) with atomic coordinates}

\item{atom_pairs}{matrix with each row having the names or indices of an atom pair (first dimension in \code{coord_array})}

\item{grouping_mat}{integer matrix (groups, models) giving groupings of models to average interaction tensors}

\item{g0}{target group norm squared values}

\item{loss_func}{loss function to use to calculate energy}

\item{...}{additional parameters passed to \code{loss_func}}

\item{gradient}{a logical value indicating whether to calculate the derivative}
}
\value{
total restraint energy calculated using \code{loss_func}

The optional derivative is contained in the \code{"gradient"} attribute. It is a 3D array (atoms, xyz, models).
}
\description{
Calculate g value restraint energy from atomic coordinates
}
