% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_array_to_r_array}
\alias{coord_array_to_r_array}
\title{Calculate internuclear vectors from atomic coordinates}
\usage{
coord_array_to_r_array(coord_array, atom_pairs)
}
\arguments{
\item{coord_array}{3D array (atoms, xyz, models) with atomic coordinates}

\item{atom_pairs}{matrix with each row having the names or indices of an atom pair (first dimension in \code{coord_array})}
}
\value{
3D array (pairs, models, xyz) with internuclear vectors. Atom pair names follow the format \code{resSeq:Atom-resSeq:Atom}.
}
\description{
Calculate internuclear vectors from atomic coordinates
}
