% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_array_to_r_array_backprop}
\alias{coord_array_to_r_array_backprop}
\title{Back-propagate energy derivative from r array to coordinates}
\usage{
coord_array_to_r_array_backprop(
  d_energy_d_coord_array,
  atom_pairs,
  d_energy_d_r_array
)
}
\arguments{
\item{d_energy_d_coord_array}{3D array (atoms, xyz, models) to accumulate derivatives into}

\item{atom_pairs}{matrix with each row having the names or indices of an atom pair (first dimension in \code{coord_array}}

\item{d_energy_d_r_array}{3D array (pairs, models, xyz)}
}
\value{
3D array (pairs, xyz, models) with d_energy_d_coord_array
}
\description{
Back-propagate energy derivative from r array to coordinates
}
