% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_array_to_relax}
\alias{coord_array_to_relax}
\title{Calculate relaxation rates from atomic coordinates}
\usage{
coord_array_to_relax(coord_array, rates, spec_den_relax_data_list)
}
\arguments{
\item{coord_array}{3D array (atoms, xyz, models) with atomic coordinates}

\item{rates}{named numeric vector with rates}

\item{spec_den_relax_data_list}{list of data for calculating relaxation rates}
}
\description{
Calculate relaxation rates from atomic coordinates
}
