% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_array_to_sigma}
\alias{coord_array_to_sigma}
\title{Calculate cross relaxation rates from atomic coordinates}
\usage{
coord_array_to_sigma(coord_array, rates, spec_den_data_list, proton_mhz)
}
\arguments{
\item{coord_array}{3D array (atoms, xyz, models) with atomic coordinates}

\item{rates}{named numeric vector with rates}

\item{spec_den_data_list}{list of data for calculating spectral density functions}

\item{proton_mhz}{spectrometer proton field strength in MHz}
}
\description{
Calculate cross relaxation rates from atomic coordinates
}
