% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_array_to_sigma_energy}
\alias{coord_array_to_sigma_energy}
\title{Calculate sigma restraint energy from atomic coordinates}
\usage{
coord_array_to_sigma_energy(
  coord_array,
  rates,
  spec_den_data_list,
  proton_mhz,
  loss_func = power_scaled_loss,
  ...,
  gradient = FALSE
)
}
\arguments{
\item{coord_array}{3D array (atoms, xyz, models) with atomic coordinates}

\item{rates}{named numeric vector with rates}

\item{spec_den_data_list}{list of data for calculating spectral density functions}

\item{proton_mhz}{spectrometer proton field strength in MHz}

\item{loss_func}{loss function to use}

\item{...}{additional parameters passed to \code{loss_func}}

\item{gradient}{a logical value indicating whether to calculate the derivative}
}
\value{
total restraint energy calculated using \code{loss_func}

The optional derivative is contained in the \code{"gradient"} attribute. It is a 3D array
(atoms, xyz, models).

Testing with \code{deriv_check} showed a slight systematic (~0.2\% underestimation) of the
gradient between two methyl groups. Perhaps there's some unaccounted correlation in
the purely additive derivative calculation?
}
\description{
Calculate sigma restraint energy from atomic coordinates
}
