% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_atomnames}
\alias{coord_atomnames}
\alias{coord_atomnames<-}
\title{Get or set four-character atom names}
\usage{
coord_atomnames(coord)

coord_atomnames(coord) <- value
}
\arguments{
\item{coord}{matrix or array with second dimension having columns 13 to 27 of the PDB ATOM records}

\item{value}{character vector with four characters per element}
}
\description{
Get or set four-character atom names
}
