% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_atomresnames}
\alias{coord_atomresnames}
\alias{coord_atomresnames<-}
\title{Get or set atom/residue name pairs}
\usage{
coord_atomresnames(coord)

coord_atomresnames(coord) <- value
}
\arguments{
\item{coord}{matrix or array with second dimension having columns 13 to 27 of the PDB ATOM records}

\item{value}{character vector with eight characters per element (1-4: atom name, 6-8: residue name)}
}
\description{
Get or set atom/residue name pairs
}
