% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_proton_groups}
\alias{coord_proton_groups}
\title{Select groups of protons that are typically detectable with unique chemical shifts}
\usage{
coord_proton_groups(
  coord_mat,
  alpha_group = FALSE,
  beta_group = FALSE,
  gamma_group = FALSE,
  delta_group = FALSE,
  epsilon_group = FALSE,
  amine_group = FALSE,
  methyl_group = TRUE,
  aromatic_group = TRUE,
  exclude_exchangeable = TRUE,
  include = character(),
  amber = FALSE
)
}
\arguments{
\item{alpha_group}{logical indicating whether to group aliphatic alpha protons}

\item{beta_group}{logical indicating whether to group aliphatic beta protons}

\item{gamma_group}{logical indicating whether to group aliphatic gamma protons}

\item{delta_group}{logical indicating whether to group aliphatic delta protons}

\item{epsilon_group}{logical indicating whether to group aliphatic epsilon protons}

\item{amine_group}{logical indicating whether to group asparagine/glutamine amine protons}

\item{methyl_group}{logical indicating whether to group methyl protons}

\item{aromatic_group}{logical indicating whether to group phenylalanine/tyrosine delta/epsilon protons}

\item{exclude_exchangeable}{logical indicating whether rapidly exchangeable protons should be excluded}

\item{include}{character vector of exchangeable protons to be explicitly included}

\item{amber}{logical indicating if names are from Amber force field}

\item{coord}{3xN matrix with a column for every atom}
}
\value{
Vector with selected protons. The names give the group the proton belongs to. The name will be the heavy atom name if all of the protons bound to that heavy atom are in the same group. Otherwise, the name will be the proton name.
}
\description{
Select groups of protons that are typically detectable with unique chemical shifts
}
