% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_rapidly_exchangeable}
\alias{coord_rapidly_exchangeable}
\title{Determine whether an atom is rapidly exchangeable at neutral pH}
\usage{
coord_rapidly_exchangeable(coord, amber = FALSE)
}
\arguments{
\item{coord}{matrix or array with second dimension having columns 13 to 27 of the PDB ATOM records}

\item{amber}{logical indicating whether atom names are from AMBER force field}
}
\description{
Determine whether an atom is rapidly exchangeable at neutral pH
}
