% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_resnames}
\alias{coord_resnames}
\alias{coord_resnames<-}
\title{Get or set three-character residue names}
\usage{
coord_resnames(coord)

coord_resnames(coord) <- value
}
\arguments{
\item{coord}{matrix or array with second dimension having columns 13 to 27 of the PDB res records}

\item{value}{character vector with three characters per element}
}
\description{
Get or set three-character residue names
}
