% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_resseq}
\alias{coord_resseq}
\alias{coord_resseq<-}
\title{Get or set residue sequence number}
\usage{
coord_resseq(coord)

coord_resseq(coord) <- value
}
\arguments{
\item{coord}{matrix or array with second dimension having columns 13 to 27 of the PDB ATOM records}

\item{value}{integer or character vector with no more than four digits}
}
\description{
Get or set residue sequence number
}
