% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{coord_to_atom_lines}
\alias{coord_to_atom_lines}
\title{Convert a matrix of coordinates into PDB formatted atom lines}
\usage{
coord_to_atom_lines(coord)
}
\arguments{
\item{coord}{3xN matrix with a column for every atom}
}
\value{
a character vector with the PDB ATOM lines
}
\description{
Make PDB ATOM lines from a 3xN matrix of coordinates
}
\details{
The names of the columns are expected to correspond to columns 13 to 27 of the PDB
atom records.
}
