% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{d_array_to_g_matrix}
\alias{d_array_to_g_matrix}
\title{Calculate group norm squared from dipole-dipole interaction tensors}
\usage{
d_array_to_g_matrix(d_array, grouping_mat, gradient = FALSE)
}
\arguments{
\item{d_array}{3D array (pairs, models, tensor elements) with interaction tensors}

\item{grouping_mat}{integer matrix (groups, models) giving groupings of models to average interaction tensors}

\item{gradient}{a logical value indicating whether to calculate the derivative}
}
\value{
vector with norm squared for each atom pair.

The optional derivative is contained in the \code{"gradient"} attribute. It is a 3D array (pairs, models, tensor elements).
}
\description{
Calculate group norm squared from dipole-dipole interaction tensors
}
