% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{d_array_to_g_matrix_backprop}
\alias{d_array_to_g_matrix_backprop}
\title{Back-propagate energy derivative from g matrix to d array}
\usage{
d_array_to_g_matrix_backprop(d_g_matrix_d_d_array, d_energy_d_g_matrix)
}
\arguments{
\item{d_g_matrix_d_d_array}{4D array (pairs, models, tensor components, groups) fom
\code{gradient} attribute of \code{d_array_to_g_matrix()}}

\item{d_energy_d_g_matrix}{matrix (pairs, groups)}
}
\value{
3D array (pairs, models, tensor elements) with d_energy_d_d_array
}
\description{
Back-propagate energy derivative from g matrix to d array
}
