% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{deriv_check}
\alias{deriv_check}
\title{Check analytical derivatives via finite difference approximation}
\usage{
deriv_check(func, value, dv, vdims, gdims, ...)
}
\arguments{
\item{func}{function to evaluate}

\item{value}{first argument to pass to \code{func}}

\item{dv}{delta used to increment elements of \code{value}}

\item{vdims}{vector of dimensions that must be incremented separately}

\item{gdims}{vector of dimensions in gradient corresponding to \code{vdims}}

\item{...}{additional arguments passed to \code{func}}
}
\value{
difference between \code{gradient} attribute returned by \code{func} and finite difference approximation
}
\description{
Check analytical derivatives via finite difference approximation
}
