% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{equiv_list_name}
\alias{equiv_list_name}
\title{Update names of equivalent atoms}
\usage{
equiv_list_name(
  equiv_list,
  restype = TRUE,
  sep = ":",
  multiatom_format = c("q", "re")
)
}
\arguments{
\item{equiv_list}{list of atomid vectors where all atomids in a given vector are equivalent}

\item{restype}{logical indicating whether to prefix with one-letter residue type}

\item{sep}{character giving separator between residue and atom names}

\item{multiatom_format}{represent multiatoms with "Q" (\code{q}) or with regular expression (\code{re}) syntax}
}
\value{
\code{equiv_list} with updated names
}
\description{
Update names of equivalent atoms
}
