% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{euler_zyz_matrix}
\alias{euler_zyz_matrix}
\title{Construct a ZYZ Euler rotation matrix}
\usage{
euler_zyz_matrix(alpha, beta, gamma)
}
\arguments{
\item{alpha}{first rotation angle in radians about the \code{z} axis}

\item{beta}{second rotation angle in radians about the \code{y} axis}

\item{gamma}{third rotation angle in radians about the \code{z} axis}
}
\value{
\verb{3 x 3} numeric rotation matrix
}
\description{
Return the \verb{3 x 3} rotation matrix corresponding to the active
\verb{Rz(alpha) Ry(beta) Rz(gamma)} Euler-angle convention.
}
\details{
This convention matches the standard parameterization used for the
rank-2 Wigner rotation matrix \code{D^(2)(alpha, beta, gamma)}.
}
\examples{
euler_zyz_matrix(0, 0, 0)
euler_zyz_matrix(pi / 2, 0, 0)

}
