% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{expand_subset_mat}
\alias{expand_subset_mat}
\title{Expand subset matrix consistent with applying Kronecker product to original rate matrix}
\usage{
expand_subset_mat(rate_subset_mat, permutation_counts)
}
\arguments{
\item{rate_subset_mat}{matrix (ensemble members, rates) giving the subset each ensemble member belongs to}

\item{permutation_counts}{vector with counts of states and names giving the rate}
}
\description{
Expand subset matrix consistent with applying Kronecker product to original rate matrix
}
\examples{
rate_mat_fast <- rate_mat_simple(10, c("f1", "f2"))
rate_mat_fast <- rate_mat_diag(rate_mat_fast)
subset_mat <- calc_subset_mat(eigen(rate_mat_fast))
expand_subset_mat(subset_mat, list("4"=2))

}
