% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{find_methyl_permutations}
\alias{find_methyl_permutations}
\title{Find permutations to coordinates of XH3 groups}
\usage{
find_methyl_permutations(atomids)
}
\arguments{
\item{atomids}{character vector of atom IDs (PDB columns 13-27)}
}
\value{
A character matrix with columns corresponding to a different permutation.
Columns correspond to equivalent atoms whose coordinates should be swapped.
}
\description{
Find permutations to coordinates of XH3 groups
}
\examples{
pdb2lum <- read_ensemble("https://files.rcsb.org/download/2LUM.pdb", proton_only=TRUE)
find_methyl_permutations(dimnames(pdb2lum)[[2]])

}
