% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{g_matrix_to_a_matrix}
\alias{g_matrix_to_a_matrix}
\title{Calculate matrix of a values from a matrix of g values}
\usage{
g_matrix_to_a_matrix(g_matrix, a_coef)
}
\arguments{
\item{g_matrix}{matrix of g values with columns associated with different groupings}

\item{a_coef}{coefficients used for calculating a values}
}
\description{
Calculate matrix of a values from a matrix of g values
}
