% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{g_matrix_to_a_matrix_backprop}
\alias{g_matrix_to_a_matrix_backprop}
\title{Back-propagate energy derivative from a matrix to g matrix}
\usage{
g_matrix_to_a_matrix_backprop(a_coef, d_energy_d_a_matrix)
}
\arguments{
\item{a_coef}{coefficients used for calculating a values}

\item{d_energy_d_a_matrix}{matrix (pairs, eigenvalues)}
}
\value{
matrix (pairs, groupings) with d_energy_d_g_matrix
}
\description{
Back-propagate energy derivative from a matrix to g matrix
}
