% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{g_to_energy}
\alias{g_to_energy}
\title{Calculate restraint energy from group norm squared values}
\usage{
g_to_energy(g, g0, k = 1, gradient = FALSE)
}
\arguments{
\item{g}{current group norm squared values}

\item{g0}{target group norm squared values}

\item{k}{force constant}

\item{gradient}{a logical value indicating whether to calculate the derivative}
}
\value{
restraint energy calculated using \eqn{k*(g-g0)^2}
}
\description{
Calculate restraint energy from group norm squared values
}
