% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{get_eigen_groups}
\alias{get_eigen_groups}
\title{Get \code{k} coefficients for creating linear combinations of \code{a} values from \code{g} values}
\usage{
get_eigen_groups(
  trans_rate_eigen,
  all_rates = NULL,
  subset_mat = NULL,
  validate = FALSE,
  eps_factor = 0.5
)
}
\arguments{
\item{trans_rate_eigen}{list with \code{values} and \code{vectors} as returned by \code{eigen()}}

\item{all_rates}{numeric vector of all rates}

\item{subset_mat}{created by \code{calc_subset_mat()} or \code{expand_subset_mat()}}

\item{validate}{check to see whether result is correct using another method}

\item{eps_factor}{epsilon used for heuristically matching rates generated in two ways}
}
\description{
Get \code{k} coefficients for creating linear combinations of \code{a} values from \code{g} values
}
\examples{
rate_mat <- rate_mat_intra_inter(c(0,0,1,1), 10, 4)
rate_mat <- rate_mat_diag(rate_mat)
get_eigen_groups(eigen(rate_mat))

}
