% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{get_rate_count_mat}
\alias{get_rate_count_mat}
\title{Get counts enabling calculation of eigenvalues from individual rates}
\usage{
get_rate_count_mat(trans_rate, all_permutations, eps_factor = 0.5)
}
\arguments{
\item{trans_rate}{transition rate matrix}

\item{all_permutations}{all possible permutations (i.e. methyl, aromatic) applied}

\item{eps_factor}{rates must differ by this amount times the lowest rate}
}
\description{
Get counts enabling calculation of eigenvalues from individual rates
}
\examples{
rate_mat <- rate_mat_intra_inter(c(0,0,1,1), 10, 4)
get_rate_count_mat(rate_mat, NULL)

}
