% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{get_rate_data}
\alias{get_rate_data}
\title{Construct \code{k} coefficients in a recursive manner}
\usage{
get_rate_data(
  trans_rate,
  parent_data = NULL,
  all_rates = NULL,
  validate = FALSE
)
}
\arguments{
\item{trans_rate}{transition rate matrix}

\item{parent_data}{previous data to modify assuming application of Kronecker product}

\item{all_rates}{vector of all rates}

\item{validate}{passed onto \code{get_eigen_groups()}}
}
\description{
Construct \code{k} coefficients in a recursive manner
}
\examples{
rate_mat_fast <- rate_mat_simple(10, c("f1", "f2"))
rate_mat_slow <- rate_mat_simple(4, c("s1", "s2"))
rate_data_base <- get_rate_data(rate_mat_fast)
rate_data <- get_rate_data(rate_mat_slow, rate_data_base)

}
