% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{get_rate_groups}
\alias{get_rate_groups}
\title{Find groups of numerically equivalent rates (degenerate eigenvalues)}
\usage{
get_rate_groups(rates, eps_factor = 0.5, eps_log10 = NULL)
}
\arguments{
\item{rates}{numeric vector of rates}

\item{eps_factor}{rates must differ by this amount times the lowest rate}

\item{eps_log10}{log10(rates) must differ by this amount

If set, eps_log10 is used and eps_factor is bypassed.}
}
\description{
Find groups of numerically equivalent rates (degenerate eigenvalues)
}
