% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{make_atom_perm_list}
\alias{make_atom_perm_list}
\title{Create list of matrices for doing two or three atom permutations}
\usage{
make_atom_perm_list(atomids, atom_permutations)
}
\arguments{
\item{atomids}{character vector of atom IDs (PDB columns 13-27)}

\item{atom_permutations}{list of permutations from \code{find_methyl_permutations()} and
\code{find_aromatic_permutations()}}
}
\description{
Create list of matrices for doing two or three atom permutations
}
\examples{
pdb2lum <- read_ensemble("https://files.rcsb.org/download/2LUM.pdb", proton_only=TRUE)
perm_methyl <- find_methyl_permutations(dimnames(pdb2lum)[[2]])
perm_aro <- find_aromatic_permutations(dimnames(pdb2lum)[[2]])
perm_list <- make_atom_perm_list(dimnames(pdb2lum)[[2]], c(perm_methyl, perm_aro))
perm_list[[3]][unlist(lapply(perm_methyl, "[", , 1)),]
perm_list[[2]][unlist(lapply(perm_aro, "[", , 1)),]

}
