% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{make_equiv_list}
\alias{make_equiv_list}
\title{Create list of equivalent atoms}
\usage{
make_equiv_list(atomids, permutations)
}
\arguments{
\item{atomids}{character vector of atom IDs (PDB columns 13-27)}

\item{permutations}{list of permutations from \code{find_methyl_permutations()} and
\code{find_aromatic_permutations()}}
}
\value{
list of atomid vectors where all atomids in a given vector are equivalent
}
\description{
Create list of equivalent atoms
}
