% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{make_ke_data}
\alias{make_ke_data}
\title{Encapsulate kinetic ensemble data into a list data structure}
\usage{
make_ke_data(
  coord_array,
  base_rate_mat,
  base_rates,
  kc,
  kmethyl = 1/1e-12,
  karo = 1/1e-04,
  proton_mhz,
  mix_times
)
}
\arguments{
\item{coord_array}{3D array (xyz, atoms, models) with atomic coordinates}

\item{base_rate_mat}{rate matrix for transitions between ensemble members}

\item{base_rates}{named vector of rates}

\item{kc}{reciprocal of molecular tumbling time}

\item{kmethyl}{methyl rotation eigenvalue}

\item{karo}{phenylalanine/tyrosine flip eigenvalue}

\item{proton_mhz}{proton frequency in MHz}

\item{mix_times}{mixing times to calculate}
}
\description{
Encapsulate kinetic ensemble data into a list data structure
}
