% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{make_spec_den_data}
\alias{make_spec_den_data}
\title{Create data structures for calculating sigma/rho from dipole-dipole interaction tensors}
\usage{
make_spec_den_data(
  ke_data,
  equiv_pair_mat,
  perm_internal = FALSE,
  sigma = NULL
)
}
\arguments{
\item{ke_data}{list structure returned by \code{make_ke_data()}}

\item{equiv_pair_mat}{matrix with pairs of equivalent atom labels matching names of
\code{ke_data$equiv_list}}

\item{perm_internal}{logical indicating whether to include atom pair internal to
permutation groups}

\item{sigma}{optional numeric vector of cross relaxation rates}
}
\description{
Create data structures for calculating sigma/rho from dipole-dipole interaction tensors
}
