% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_traj.R
\name{noe_dacf_kernel}
\alias{noe_dacf_kernel}
\title{Calculate kernel weights for determining NOE sigma and rho from a dipolar autocorrelation function}
\usage{
noe_dacf_kernel(
  dacf,
  dt,
  tauc,
  proton_mhz,
  terms = c("0", "omega", "2omega"),
  ntrunc = 1
)
}
\arguments{
\item{dacf}{example dipolar autocorrelation function vector or the length of the vector}

\item{dt}{time interval between dacf elements in seconds}

\item{tauc}{isotropic tumbling time of the molecule in seconds}

\item{proton_mhz}{proton Larmor frequency in MHz}

\item{terms}{vector of terms to use in calculating the kernel}

\item{ntrunc}{number of trailing kernel elements over which to spread the truncated tail contribution}
}
\description{
Calculate kernel weights for determining NOE sigma and rho from a dipolar autocorrelation function
}
