% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_traj.R
\name{pdb_traj_acf}
\alias{pdb_traj_acf}
\title{Calculate a dipole-dipole interaction tensor autocorrelation function}
\usage{
pdb_traj_acf(coord_buffer, nframes, segment_nums, atom_pair_mat_list, acf_dir)
}
\arguments{
\item{coord_buffer}{3D array (frames, xyz, atoms) with nframes*(2+length(segment_nums)) frames}

\item{nframes}{number of trajectory frames per segment}

\item{segment_nums}{segment numbers for the middle of the buffer}

\item{atom_pair_mat_list}{list of matrices (pairs, 2)}

\item{acf_dir}{directory where <segment_num>.rds files written with calculated ACF}
}
\value{
list with NULL values of length segment_nums
}
\description{
Calculate a dipole-dipole interaction tensor autocorrelation function
}
