% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_traj.R
\name{pdb_traj_apply}
\alias{pdb_traj_apply}
\title{Apply a function to a rolling buffer of trajectory frames}
\usage{
pdb_traj_apply(traj_data, FUN, nframes = 1000, nsegments = 10, skip = 1, ...)
}
\arguments{
\item{traj_data}{list structure with data necessary for reads}

\item{FUN}{function that takes parameters coord_buffer, nframes, and segment_nums and returns a list with element for each segment}

\item{nframes}{number of trajectory frames per segment}

\item{nsegments}{number of segments passed to FUN}

\item{skip}{number of frames to skip that are included in the first buffer region}

\item{...}{additional arguments passed to FUN}
}
\value{
list concatenating returns of all the FUN calls
}
\description{
Apply a function to a rolling buffer of trajectory frames
}
