% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_traj.R
\name{pdb_traj_close}
\alias{pdb_traj_close}
\title{Close a PDB trajectory file}
\usage{
pdb_traj_close(traj_data)
}
\arguments{
\item{traj_data}{list structure with data necessary for reads}
}
\description{
Close a PDB trajectory file
}
