% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_traj.R
\name{pdb_traj_d}
\alias{pdb_traj_d}
\title{Calculate second-rank dipole interaction tensors}
\usage{
pdb_traj_d(
  coord_buffer,
  nframes,
  segment_nums,
  atom_pair_mat_list,
  unit = FALSE,
  d_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{coord_buffer}{3D array (frames, xyz, atoms) with nframes*(2+length(segment_nums)) frames}

\item{nframes}{number of trajectory frames per segment}

\item{segment_nums}{segment numbers for the middle of the buffer}

\item{atom_pair_mat_list}{list of matrices (pairs, 2)}

\item{unit}{logical indicating whether to convert to unit vectors}

\item{d_dir}{directory where <segment_num>.rds files written with calculated tensors}

\item{verbose}{logical indicating whether to emit progress messages}
}
\value{
list with NULL values of length segment_nums
}
\description{
Calculate second-rank dipole interaction tensors
}
