% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_traj.R
\name{pdb_traj_dacf}
\alias{pdb_traj_dacf}
\title{Calculate second-rank dipolar autocorrelation function}
\usage{
pdb_traj_dacf(
  coord_buffer,
  nframes,
  segment_nums,
  atom_pair_mat_list,
  dacf_dir = NULL
)
}
\arguments{
\item{coord_buffer}{3D array (frames, xyz, atoms) with nframes*(2+length(segment_nums)) frames}

\item{nframes}{number of trajectory frames per segment}

\item{segment_nums}{segment numbers for the middle of the buffer}

\item{atom_pair_mat_list}{list of matrices (pairs, 2)}

\item{dacf_dir}{directory where <segment_num>.rds files written with calculated DACF}
}
\value{
list with NULL values of length segment_nums
}
\description{
Calculate second-rank dipolar autocorrelation function
}
