% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_traj.R
\name{pdb_traj_open}
\alias{pdb_traj_open}
\title{Open a PDB trajectory file for reading}
\usage{
pdb_traj_open(filename, atomids = NULL, frame_max_lines = 10000)
}
\arguments{
\item{filename}{file name of \code{.pdb} or \code{.pdb.gz}}

\item{atomids}{character vector of atom coordinates to return matching characters 13-27 of ATOM records}

\item{frame_max_lines}{maximum lines to read that must include the first frame}
}
\value{
list structure with data necessary for reads
}
\description{
Open a PDB trajectory file for reading
}
