% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdb_traj.R
\name{pdb_traj_read}
\alias{pdb_traj_read}
\title{Read coordinates from the trajectory}
\usage{
pdb_traj_read(traj_data, nframes = 1000, check_atomids = TRUE)
}
\arguments{
\item{traj_data}{list structure with data necessary for reads}

\item{nframes}{number of frames to read from the trajectory}

\item{check_atomids}{check whether read atomids equal those expected (taking a performance hit)}
}
\value{
3D array (xyz, frames, atoms) with atomic coordinates
}
\description{
Read coordinates from the trajectory
}
