% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{power_scale}
\alias{power_scale}
\title{Scale real or back calculated values towards zero by taking a power}
\usage{
power_scale(x, p)
}
\arguments{
\item{x}{numerical values to be rescaled}

\item{p}{numeric power to raise the values to, usually 1 or less}
}
\value{
\deqn{ ( ( |x| + 1 )^p - 1 ) \operatorname{sgn}(x) }
}
\description{
Scale real or back calculated values towards zero by taking a power
}
