% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{power_scaled_loss}
\alias{power_scaled_loss}
\title{Traditional squared loss function scaled by a power}
\usage{
power_scaled_loss(x, x0, p = 1, k = 1, gradient = FALSE)
}
\arguments{
\item{x}{current values}

\item{x0}{target values}

\item{p}{numeric power to raise the values to, usually 1 or less}

\item{k}{force constant}

\item{gradient}{a logical value indicating whether to calculate the derivative}
}
\value{
\deqn{ k \left ( ( ( |x| + 1 )^p - 1 ) \operatorname{sgn}(x) - ( ( |x_0| + 1 )^p - 1 ) \operatorname{sgn}(x_0) \right )^2 }
}
\description{
Traditional squared loss function scaled by a power
}
\examples{
x <- seq(-10, 10, by=0.1)
loss_x <- sapply(x, power_scaled_loss, 2, 0.25)
loss_x_grad <- power_scaled_loss(x, 2, 0.25, gradient=TRUE)
par(mfrow=c(2, 1))
plot(x, loss_x, type="l", ylab="loss")
plot(x, attr(loss_x_grad, "gradient"), type="l", ylab="dloss/dx")
points(x[-1]-mean(diff(x))/2, diff(loss_x)/mean(diff(x)), type="l", col="blue")
abline(h = 0, col="gray")
legend("bottomright", legend=c("Analytical", "Finite Difference"), bty="n", lwd=1, col=c("black", "blue"))

}
