% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{quadratic_loss}
\alias{quadratic_loss}
\title{Quadratic loss function}
\usage{
quadratic_loss(x, x0, k = 1, gradient = FALSE)
}
\arguments{
\item{x}{current values}

\item{x0}{target values}

\item{k}{force constant}

\item{gradient}{a logical value indicating whether to calculate the derivative}
}
\value{
\deqn{k(x - x_0)^2}
}
\description{
Quadratic loss function
}
