% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{r_array_to_d_array_backprop}
\alias{r_array_to_d_array_backprop}
\title{Back-propagate energy derivative from d array to r array}
\usage{
r_array_to_d_array_backprop(d_d_array_d_r_array, d_energy_d_d_array)
}
\arguments{
\item{d_d_array_d_r_array}{4D array (pairs, models, tensor elements, xyz) from
\code{gradient} attribute of \code{r_array_to_d_array()}}

\item{d_energy_d_d_array}{3D array (pairs, models, tensor elements)}
}
\value{
3D array (pairs, models, xyz) with d_energy_d_r_array
}
\description{
Back-propagate energy derivative from d array to r array
}
