% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{rate_data_to_mat_list}
\alias{rate_data_to_mat_list}
\title{Convert rate data into a format more easily serializable to set of tables}
\usage{
rate_data_to_mat_list(rate_data, rate_count_mat, rates_named = NULL)
}
\arguments{
\item{rate_data}{list returned by \code{get_rate_data()}}

\item{rate_count_mat}{matrix returned by \code{get_rate_count_mat()} including methyl and
aromatic rates}

\item{rates_named}{numeric vector whose names give the rates to be used}
}
\value{
a list with elements: \code{groupings}, \code{a_coef}, and \code{lambda_coef}
}
\description{
Convert rate data into a format more easily serializable to set of tables
}
