% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{rate_mat_diag}
\alias{rate_mat_diag}
\title{Update transition rate diagonal using rates}
\usage{
rate_mat_diag(rate_mat)
}
\arguments{
\item{rate_mat}{square transition rate matrix}
}
\description{
Update transition rate diagonal using rates
}
