% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{rate_mat_intra_inter}
\alias{rate_mat_intra_inter}
\title{Create hierarchical transition rate matrix}
\usage{
rate_mat_intra_inter(group_vec, k_intra, k_inter)
}
\arguments{
\item{group_vec}{integer vector giving group membership of each state}

\item{k_intra}{desired eigenvalue for intra-group transitions}

\item{k_inter}{desired eigenvalue for inter-group transitions}
}
\description{
Create hierarchical transition rate matrix
}
\examples{
rate_mat <- rate_mat_intra_inter(c(0,0,1,1), 10, 4)
rate_mat <- rate_mat_diag(rate_mat)
eigen(rate_mat)

}
