% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{rate_mat_kronecker}
\alias{rate_mat_kronecker}
\title{Calculate Kronecker product of two transition rate matrices}
\usage{
rate_mat_kronecker(rate_mat_a, rate_mat_b)
}
\arguments{
\item{rate_mat_a}{first transition rate matrix}

\item{rate_mat_b}{second transition rate matrix}
}
\description{
Calculate Kronecker product of two transition rate matrices
}
\examples{
rate_mat_fast <- rate_mat_simple(10, c("f1", "f2"))
rate_mat_slow <- rate_mat_simple(4, c("s1", "s2"))
rate_mat <- rate_mat_kronecker(rate_mat_fast, rate_mat_slow)
rate_mat <- rate_mat_diag(rate_mat)
eigen(rate_mat)

}
