% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{rate_mat_simple}
\alias{rate_mat_simple}
\title{Create transition rate matrix with single rate}
\usage{
rate_mat_simple(k, n_names)
}
\arguments{
\item{k}{desired eigenvalue for the rate matrix}

\item{n_names}{character vector with row and columns}
}
\description{
Create transition rate matrix with single rate
}
\examples{
rate_mat <- rate_mat_simple(4, c("a", "b", "c", "d"))
rate_mat <- rate_mat_diag(rate_mat)
eigen(rate_mat)

}
