% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{read_ensemble}
\alias{read_ensemble}
\title{Read an ensemble of PDB files}
\usage{
read_ensemble(pdb_files, model_idx = NULL, proton_only = FALSE)
}
\arguments{
\item{pdb_files}{character vector paths to PDB format file(s)}

\item{model_idx}{vector with indices of models to read}

\item{proton_only}{logical indicating whether to only read hydrogen ATOM records}
}
\value{
Numeric coordinate array. For a single input file the dimensions are
\verb{(xyz, atoms, models)}; for multiple input files an additional trailing file
dimension is retained.
}
\description{
Read an ensemble of PDB files
}
