% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{read_spec_den_data}
\alias{read_spec_den_data}
\title{Read data for calculating spectral density functions}
\usage{
read_spec_den_data(prefix_path)
}
\arguments{
\item{prefix_path}{to prefix of four CSV files}
}
\value{
a list with elements: \code{atom_pairs}, \code{groupings}, \code{a_coef}, and \code{lambda_coef}
}
\description{
Read data for calculating spectral density functions
}
