% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\docType{data}
\name{aa1_to_aa3}
\alias{aa1_to_aa3}
\alias{aa3_to_aa1}
\title{Convert between one and three letter residue names}
\format{
An object of class \code{character} of length 20.

An object of class \code{character} of length 20.
}
\usage{
aa1_to_aa3[aa1]

aa3_to_aa1[aa3]
}
\description{
Convert between one and three letter residue names
}
\keyword{datasets}
