% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{subset_idx_to_grouping_mat}
\alias{subset_idx_to_grouping_mat}
\title{Convert a list of subset groups into a matrix with group indices}
\usage{
subset_idx_to_grouping_mat(subset_idx)
}
\arguments{
\item{subset_idx}{list of lists of integers giving members of each group}
}
\value{
a matrix with row for each outer list and group assignment for each ensemble member
}
\description{
Convert a list of subset groups into a matrix with group indices
}
