% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{test_eros3}
\alias{test_eros3}
\title{Tests with EROS3 ensemble subset (First two ensemble members and five atoms from M1)}
\usage{
test_eros3()
}
\description{
Tests with EROS3 ensemble subset (First two ensemble members and five atoms from M1)
}
