% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{test_random}
\alias{test_random}
\title{Tests with random internuclear vectors of length 1}
\usage{
test_random()
}
\value{
No return value, called for side effects.
}
\description{
Tests with random internuclear vectors of length 1
}
