% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{test_toy}
\alias{test_toy}
\title{Tests with toy example from Smith 2020 J Biomol NMR}
\usage{
test_toy()
}
\value{
No return value, called for side effects.
}
\description{
Tests with toy example from Smith 2020 J Biomol NMR
}
