% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke_prep.R
\name{unique_atom_pair_map}
\alias{unique_atom_pair_map}
\title{Generate matrix of possible permuted assignments of atom pairs within a group}
\usage{
unique_atom_pair_map(permutation)
}
\arguments{
\item{permutation}{single permutation matrix

This is useful for calculating interactions within a permutation group}
}
\description{
Generate matrix of possible permuted assignments of atom pairs within a group
}
\examples{
pdb2lum <- read_ensemble("https://files.rcsb.org/download/2LUM.pdb", proton_only=TRUE)
perm_methyl <- find_methyl_permutations(dimnames(pdb2lum)[[2]])
perm_aro <- find_aromatic_permutations(dimnames(pdb2lum)[[2]])
unique_atom_pair_map(perm_methyl[[1]])
unique_atom_pair_map(perm_aro[[1]])

}
