% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\name{a_matrix_to_sigma}
\alias{a_matrix_to_sigma}
\title{Calculate sigma from a matrix of a values}
\usage{
a_matrix_to_sigma(a_matrix, lambda_prime_vec, proton_mhz, gradient = FALSE)
}
\arguments{
\item{a_matrix}{matrix of a values with columns associated with eigenvalues}

\item{lambda_prime_vec}{eigenvalues augmented with tumbling rate}

\item{proton_mhz}{spectrometer proton field strength in MHz}

\item{gradient}{a logical value indicating whether to calculate the derivative}
}
\description{
Calculate sigma from a matrix of a values
}
