% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ke.R
\docType{data}
\name{amber_to_pdb}
\alias{amber_to_pdb}
\alias{pdb_to_amber}
\title{Convert between proton atom/residue name pairs}
\format{
An object of class \code{character} of length 160.

An object of class \code{character} of length 160.
}
\usage{
amber_to_pdb

pdb_to_amber
}
\description{
Convert between proton atom/residue name pairs
}
\keyword{datasets}
